step2/requirements.txt
azure-functions==1.11.2
fastapi==0.85.0
nest-asyncio==1.5.5
